AddAssignments, individual course page
CoursePage.java:
Add onClick response on list view
IndividualCourse.java:
Able to show the initial data with course title and units (PNP/LETTER
not yet). Need to know if we need a navigation drawer on it, or a
go-back button on it.
AddAssignments.java:
Able to add assignments and show on IndividualCourse page. NEED TO
CHECK EXCEPTIONS
courses.java:
Since we don’t have a back-end data storage thing, I created a class
and store the assignments, units, and course titles, and PNP options in
it. Didn’t test for whether showing the correct course title using this
class since I used an array list before. NEED TO TEST.